Replication Experiment: Continuous integration in a social-coding world  Empirical evidence from GITHUB
The social coding world is an approach to software engineering mainly
based on the Git version control system. In the last years these code
sharing platforms had an enormous success, thanks to how easy is to
share code and the usefulness of having another point of view of the
problem from the online community. With this enormous number of
dislocated developers the quality of the code written can get worse and
the testing and building process can take a lot of time to get done. One
of the most used on GitHub is Travis CI that detects automatically when
a commit or a pull request is pushed to any branch of the repository
that uses Travis CI. In this replication experiment we will re examin
how the type of contribution and the predisposition to run automated
build and tests can affect the successfulness of a project.